requirements.txt
psycopg[binary]==3.2.3
python-dotenv==1.0.1
rich==13.9.4
